+++ src/modules/admin/views/redirect/create.php
@@ -1,7 +1,6 @@
 <?php
 /**
- * Create redirect form.
- * @see davidhirtz\yii2\skeleton\modules\admin\controllers\RedirectController::actionCreate()
+ * @see RedirectController::actionCreate()
  *
  * @var View $this
  * @var Redirect $redirect
@@ -9,6 +8,8 @@
 
 use davidhirtz\yii2\skeleton\helpers\Html;
 use davidhirtz\yii2\skeleton\models\Redirect;
+use davidhirtz\yii2\skeleton\modules\admin\controllers\RedirectController;
+use davidhirtz\yii2\skeleton\modules\admin\widgets\forms\RedirectActiveForm;
 use davidhirtz\yii2\skeleton\web\View;
 use davidhirtz\yii2\skeleton\widgets\bootstrap\Panel;
 use yii\helpers\Url;
@@ -24,7 +25,7 @@ $this->setBreadcrumb(Yii::t('skeleton', 'Redirects'), ['index']);
 
 <?= Panel::widget([
     'title' => $this->title,
-    'content' => $redirect->getActiveForm()::widget([
+    'content' => RedirectActiveForm::widget([
         'model' => $redirect,
     ]),
 ]);