+++ src/modules/admin/views/redirect/update.php
@@ -1,7 +1,6 @@
 <?php
 /**
- * Update redirect form.
- * @see davidhirtz\yii2\skeleton\modules\admin\controllers\RedirectController::actionUpdate()
+ * @see RedirectController::actionUpdate()
  *
  * @var View $this
  * @var Redirect $redirect
@@ -10,7 +9,9 @@
 
 use davidhirtz\yii2\skeleton\helpers\Html;
 use davidhirtz\yii2\skeleton\models\Redirect;
+use davidhirtz\yii2\skeleton\modules\admin\controllers\RedirectController;
 use davidhirtz\yii2\skeleton\modules\admin\data\RedirectActiveDataProvider;
+use davidhirtz\yii2\skeleton\modules\admin\widgets\forms\RedirectActiveForm;
 use davidhirtz\yii2\skeleton\modules\admin\widgets\grids\RedirectGridView;
 use davidhirtz\yii2\skeleton\web\View;
 use davidhirtz\yii2\skeleton\widgets\bootstrap\Panel;
@@ -28,11 +29,10 @@ $this->setBreadcrumb(Yii::t('skeleton', 'Redirects'), ['index']);
 
 <?= Panel::widget([
     'title' => $this->title,
-    'content' => $redirect->getActiveForm()::widget([
+    'content' => RedirectActiveForm::widget([
         'model' => $redirect,
     ]),
-]);
-?>
+]); ?>
 
 <?= Panel::widget([
     'title' => Yii::t('skeleton', 'Additional Redirects'),